iomd button latchy thing is 32bit wide and also contains setting for 16bit sound.
--HG--
branch : trunk

+++ src/iomd.h
@@ -145,7 +145,7 @@ extern void iomd_reset(IOMDType type);
 extern void iomd_end(void);
 extern uint32_t iomd_read(uint32_t addr);
 extern void iomd_write(uint32_t addr, uint32_t val);
-extern uint8_t iomd_mouse_buttons_read(void);
+extern uint32_t iomd_mouse_buttons_read(void);
 extern void iomd_vsync(int vsync);
 
 extern void gentimerirq(void);
